fixed flex direction for desktop and moved render to rendering-monie-cards.js
